refactor(AresCachedResults): rename file to match class name
