+++ src/lib/classes/cachedResults.ts
