+++ src/lib/classes/index.ts
@@ -2,7 +2,7 @@ export * from "./base";
 export * from "./baseManager";
 export * from "./baseModule";
 export * from "./cachedManager";
-export * from "./cacheResults";
+export * from "./cachedResults";
 export * from "./client";
 export * from "./error";
 export * from "./baseResults";
